requirements.txt
APScheduler~=3.10.4
Django~=4.2.7
djangorestframework~=3.14.0
plotly~=5.18.0
requests~=2.31.0
beautifulsoup4~=4.12.2
pandas~=2.1.3
martor~=1.6.28
gunicorn~=21.2.0
